refactor: reorganize all strategies
