+++ config/front/services.yml
@@ -24,33 +24,33 @@ services:
       - "@PrestaShop\\Module\\Psgdpr\\Repository\\LoggerRepository"
       - "@PrestaShop\\Module\\Psgdpr\\Repository\\CustomerRepository"
 
-  PrestaShop\Module\Psgdpr\Service\ExportCustomerData\ExportCustomerDataService:
-    class: 'PrestaShop\Module\Psgdpr\Service\ExportCustomerData\ExportCustomerDataService'
+  PrestaShop\Module\Psgdpr\Service\ExportService:
+    class: 'PrestaShop\Module\Psgdpr\Service\ExportService'
     arguments:
       - '@=service("prestashop.adapter.legacy.context").getContext()'
       - '@=service("Psgdpr").getTranslator()'
 
-  PrestaShop\Module\Psgdpr\Service\ExportCustomerData\ExportCustomerDataFactory:
-    class: 'PrestaShop\Module\Psgdpr\Service\ExportCustomerData\ExportCustomerDataFactory'
+  PrestaShop\Module\Psgdpr\Service\Export\ExportFactory:
+    class: 'PrestaShop\Module\Psgdpr\Service\Export\ExportFactory'
     arguments:
       - !tagged_iterator psgdpr.export.customerData
 
-  PrestaShop\Module\Psgdpr\Service\ExportCustomerData\ExportCustomerDataToCsv:
-    class: 'PrestaShop\Module\Psgdpr\Service\ExportCustomerData\ExportCustomerDataToCsv'
+  PrestaShop\Module\Psgdpr\Service\Export\Strategy\ExportToCsv:
+    class: 'PrestaShop\Module\Psgdpr\Service\Export\Strategy\ExportToCsv'
     tags: ["psgdpr.export.customerData"]
     arguments:
       - '@=service("prestashop.adapter.legacy.context").getContext()'
       - '@PrestaShop\Module\Psgdpr\Service\LoggerService'
 
-  PrestaShop\Module\Psgdpr\Service\ExportCustomerData\ExportCustomerDataToPdf:
-    class: 'PrestaShop\Module\Psgdpr\Service\ExportCustomerData\ExportCustomerDataToPdf'
+  PrestaShop\Module\Psgdpr\Service\Export\Strategy\ExportToPdf:
+    class: 'PrestaShop\Module\Psgdpr\Service\Export\Strategy\ExportToPdf'
     tags: ["psgdpr.export.customerData"]
     arguments:
       - '@=service("prestashop.adapter.legacy.context").getContext()'
       - '@PrestaShop\Module\Psgdpr\Service\LoggerService'
 
-  PrestaShop\Module\Psgdpr\Service\ExportCustomerData\ExportCustomerDataToJson:
-    class: 'PrestaShop\Module\Psgdpr\Service\ExportCustomerData\ExportCustomerDataToJson'
+  PrestaShop\Module\Psgdpr\Service\Export\Strategy\ExportToJson:
+    class: 'PrestaShop\Module\Psgdpr\Service\Export\Strategy\ExportToJson'
     tags: ["psgdpr.export.customerData"]
     arguments:
       - '@=service("prestashop.adapter.legacy.context").getContext()'
